Check if user is author of the post, in order to update or delete it

+++ app/Http/Controllers/ApiControllers/PostController.php
@@ -8,6 +8,7 @@ use App\Post;
 use App\Http\Resources\Post as PostResource;
 use App\Http\Requests\PostCreate;
 use App\Http\Requests\PostUpdate;
+use Auth;
 
 class PostController extends Controller
 {
@@ -20,7 +21,9 @@ class PostController extends Controller
     }
 
     public function store(PostCreate $request){
+        $user_id = Auth::user()->id;
         $validated = $request->validated();
+        $validated['user_id'] = $user_id;
         $post = Post::create($validated);
         return response([
             'message' => 'Post Created Successfully!',
@@ -31,15 +34,23 @@ class PostController extends Controller
     public function update($id, PostUpdate $request){
         $validated = $request->validated();
         $post = Post::findOrFail($id);
-        $post->update($validated);
-        return response([
-            'message' => 'Post has been updated',
-            'post' => new PostResource($post)
-        ]);
+        if (Auth::user()->id != $post->user->id){
+            return response(['message' => 'You are not author of this post. You can\'t edit it']);
+        }
+        else{
+            $post->update($validated);
+            return response([
+                'message' => 'Post has been updated',
+                'post' => new PostResource($post)
+            ]);
+        }
     }
 
-    public function destroy($id){
+    public function destroy($id, Request $request){
         $post = Post::findOrFail($id);
+        if (Auth::user()->id != $post->user->id){
+            return response(['message' => 'You are not author of this post. You can\'t delete it']);
+        }
         $post->delete();
         return response([
             'message' => 'Post has been deleted'
